Removed QueryCriteria since is not used
