+++ src/main/java/tech/jhipster/sample/config/DatabaseConfiguration.java
@@ -31,11 +31,14 @@ import org.springframework.util.StringUtils;
 import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
 import tech.jhipster.config.JHipsterConstants;
 import tech.jhipster.config.JHipsterProperties;
-import tech.jhipster.sample.config.springdata.CustomCouchbaseRepositoryFactoryBean;
+import tech.jhipster.sample.config.sdc.CustomCouchbaseRepositoryFactoryBean;
 
 @Configuration
 @Profile("!" + JHipsterConstants.SPRING_PROFILE_CLOUD)
-@EnableCouchbaseRepositories(basePackages = "tech.jhipster.sample.repository", repositoryFactoryBeanClass = CustomCouchbaseRepositoryFactoryBean.class)
+@EnableCouchbaseRepositories(
+    basePackages = "tech.jhipster.sample.repository",
+    repositoryFactoryBeanClass = CustomCouchbaseRepositoryFactoryBean.class
+)
 @EnableCouchbaseAuditing(auditorAwareRef = "springSecurityAuditorAware", dateTimeProviderRef = "")
 public class DatabaseConfiguration extends AbstractCouchbaseConfiguration {
