+++ src/main/java/tech/jhipster/sample/config/sdc/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Classes in this package are meant to extend Spring Data Couchbase and fix some of the issues encountered during
+ * integration.
+ */
+package tech.jhipster.sample.config.sdc;
